pyproject.toml
[project]
name = "nagara"
version = "0.1.0"
description = ""
authors = [
    { name = "Oluwadamilola Ifedayo", email = "dami.func@gmail.com" }
]
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "email-validator>=2.3.0",
    "fastapi>=0.136.0",
    "psycopg2-binary>=2.9.12",
    "pwdlib[argon2]>=0.3.0",
    "pydantic-settings>=2.14.0",
    "pyjwt>=2.12.1",
    "sqlalchemy[asyncio]>=2.0.49",
    "uvicorn[standard]>=0.44.0",
]

[project.optional-dependencies]
# Secret backends — install only the one(s) you use.
# `pip install nagara[infisical]` pulls the Infisical SDK.
infisical = ["infisicalsdk>=1.0.0"]

[build-system]
requires = ["uv_build>=0.11.7,<0.12.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "aiosqlite>=0.22.1",
    "httpx>=0.28.1",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.11",
    "ty>=0.0.32",
]

[tool.ruff]
line-length = 100
target-version = "py314"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "UP", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
